+++ src/main/java/com/kruemblegard/client/KruemblegardRenderLayers.java
@@ -2,7 +2,6 @@ package com.kruemblegard.client;
 
 import com.kruemblegard.Kruemblegard;
 import com.kruemblegard.block.AshspireCactusBlock;
-import com.kruemblegard.block.AshspireColossusBlock;
 import com.kruemblegard.block.AshveilBlock;
 import com.kruemblegard.block.KruemblegardLeavesBlock;
 import com.kruemblegard.block.WaylilyBlock;
@@ -48,7 +47,7 @@ public final class KruemblegardRenderLayers {
                         || block == ModBlocks.WAYLILY_STALK.get()
                         || block == ModBlocks.WAYLILY_STALK_BASE.get()) {
                     ItemBlockRenderTypes.setRenderLayer(block, RenderType.cutout());
-                } else if (block instanceof AshspireCactusBlock || block instanceof AshspireColossusBlock) {
+                } else if (block instanceof AshspireCactusBlock) {
                     // Match vanilla cactus: cutout for transparent pixels.
                     ItemBlockRenderTypes.setRenderLayer(block, RenderType.cutout());
                 } else if (block instanceof BushBlock) {